I want to guarantee that the symfony_env variable is always lowercased so that I can safely use. A cool way to do this is with a 'pre task'.
Add a new key called... well... pre_tasks and a new task called 'Convert entered Symfony environment to lowercase'

Pre-tasks are exactly like tasks... they just run first.

Wait... then what the heck is the difference between a 'pre task' and just putting the task at the top of the task section?

Nothing! Well, nothing yet. But later, when we talk about roles - ooh roles are fancy - there will be a difference: pre tasks run before roles.

Anywho, in this pre task, we basically want to re-set the symfony_env variable to a lowercased version of itself. To do that, we'll use a new module: set_fact. We already know that we can set variables in 3 different ways: with vars, vars_prompt or by using the register key below a task. The set_fact module is yet another way.

acts versus Variables

But wait... why set_fact and not set_variable? So... here's the deal: you'll hear the words 'facts' and 'variables' in Ansible... basically interchangeably. Both facts and variables can be set, and are referenced in exactly the same way. And while there do seem to be some subtle differences between the two, if you think of a fact and a variable as the same thing, it'll make your life easier. When you run the playbook, the first task is called setup, and it prepares some facts about the host machine. That's usually where you hear the word facts: it's info about each host.

So, we're using set_fact... to set a fact... or a variable... Set symfony_env to {{ symfony_env|lower }}
Love it! Oh, and just like with tasks, these can be tagged. We'll probably want this to run all the time... so let's use a special tag called always

Try the playbook!

ansible-playbook ansible/playbook.yml -i ansible/hosts.ini

Be difficult and use PROD in all capital letters again.

Ha, ha! This time, our playbook outsmarts us and lowercases the value.