Environment Variables and Pre Tasks
Remove the two debug tasks. Can we also remove the |lower from the environment variable? Actually... no! The environment section runs before tasks, even pre_tasks. And this has nothing to do with the order we have things in this file - environment always runs first. So, keep the |lower.

Using the symfony_env Variable

Ok, time to use our fancy symfony_env variable! First, when we install the composer dependencies, we currently have no_dev: no. But now, if the environment is prod, this can be yes. Let's use a fancy expression! {{ 'yes' if (prod == symfony_env) else 'no' }}. Don't forget your curly braces. Weird, but cool! This is a special Jinja syntax.

Conditionally Running Tasks

Next, find the fixtures task. Hmm. If we're deploying in the prod environment... we might not want to load the data fixtures at all. But so far, we don't have any way to conditionally run a task: tasks always run.

Well guess what? We can tell a task to not run with the when key. In this case, say when: 'symfony_env != 'prod'. Finally, down in the Clear cache task, instead of prod, use {{ symfony_env }}

Let's try this thing! Re-run the playbook, but use -t deploy:

ansible-playbook ansible/playbook.yml -i ansible/hosts.ini -t deploy

Use the prod environment. As we just saw, the vars_prompt runs always: it doesn't need a tag. Then, our 'pre task' should run thanks to the always tag

By the time the 'Composer Install' task executes, it should run with no_dev: yes, and then hopefully it'll skip data fixtures and change 'Clear cache'

The 'Install Composer's dependencies' does show as changed: that's a good sign: it should have installed less packages than before. And yea! It's skipping the fixtures!

In the VM, try to ls vendor/sensio. Ok cool! One of the require-dev dependencies is sensio/generator-bundle. That is not here, proving that the dev dependencies did NOT install this time. We are in business!

And before we continue, under the 'Clear Cache' task, add changed_when: false

That's not critical, it'll just prevent it from showing up as changed on every run.

Now, let's create a faster, smarter playbook by skipping some redundant tasks!