If you executed the playbook twice in a row, you would see 'Changed 2'. On every playbook run, we're doing unnecessary work: we're always downloading Composer and then moving it globally... even if it's already there!
That's not a big deal... but it is wasteful. Let's make our playbook smarter by running tasks conditionally, based on facts that we collect about the host. In other words: let's not download Composer if it already exists!

Checking if a File Exists: stat

So that's the first mission: figure out if the /usr/local/bin/composer file exists. To do that, we can use a really handy module called stat. It's similar to the Unix stat command... which gives you a ton of info about a file.

Before we download composer, add a new task called 'Check for Composer'. Use the stat module and set the path to /usr/local/bin/composer. Then, register a new variable called composer_stat. And don't forget the deploy tag! To see what goodies that variable has in it, add a debug task to print composer_stat. Give that the deploy tag too.

Ok! Change over to your terminal and run the playbook with -t deploy:

ansible-playbook ansible/playbook.yml -i ansible/hosts.ini -t deploy

There it is! Hit CTRL+C to quit.

The new variable has a stat key with a lot of info. The key we want is exists.