Remove the debug task.
Skipping Tasks

Our goal is to skip the next three tasks if that file exists. Like before, use the when key set to not composer_stat.stat.exists

In other words, check the stat.exists key, and if that is true, we want to not run this.

Copy that and put it below 'Move Composer globally', and also 'Set Permissions on Composer'

I think we're ready! Try it:

ansible-playbook ansible/playbook.yml -i ansible/hosts.ini -t deploy

Wait... yes! Skipped! Stop the playbook.