Upgrading Composer
This is a nice performance improvement... but like with a lot of things in programming, new features mean new complexity. Thanks to this... the Composer executable will eventually become really old and out-of-date. We need to make sure it's upgraded to the latest version.

Below the set permissions task, add a new one called 'Make sure Composer is at its latest version'. This time we can use the composer module. Set working_dir to {{ symfony_root_dir }}... though that doesn't matter in this case... because we'll use self-update to upgrade Composer. We can even add when: composer_stat.stat.exists to avoid running this if Composer was just downloaded. Give it our favorite deploy tag.

Ok, run it! But this time, add a --verbose flag:

ansible-playbook ansible/playbook.yml -i ansible/hosts.ini -t deploy --verbose

This is a sweet trick... because... as you can see, it prints the output from each task... which can save us from needing to add so many debug tasks to print variables.

Boom! The new Composer upgrade task ran! I'll stop with 'CTRL+C'. And, we can see the stdout: You are already using composer version 1.6.4.

So it worked! But like with other tasks... it's reporting as 'Changed' even though it did not actually change anything! We already know how to fix that.