Scroll down to the 'Create DB' task and copy its changed_when. Scroll back up and paste it under this task. Register a new variable - composer_self_update - and use that in changed_when. Then, all we need to do is search for You are already using composer version.
If that shows up in the command, then we know nothing changed. Paste that into the search filter.

Test it out!

ansible-playbook ansible/playbook.yml -i ansible/hosts.ini -t deploy --verbose

There it is! This time it's Okay - not changed. And on that note, we can even start skipping tasks based on whether or not other tasks are 'changed' or 'ok'. Let's try that next!